load gltf materials once and wrap in Rc
